@@ -4,7 +4,7 @@ This project has been licensed under the MIT license.
 Copyright (c) 2025-present pygame-community
 """
 
-__title__ = "eoc.lib"
+__title__ = "eoclib"
 __author__ = "pygame-community"
 __license__ = "MIT"
 __copyright__ = "Copyright 2025-present pygame-community"
