@@ -2,46 +2,32 @@ from collections.abc import Generator
 from typing import Callable
 
 import pygame
-from eoclib.callbacks import EOCWareMainloop, EOCWareSetup, EOCWareThumbnail
-from eoclib.types import WareConfig, WareData
+from warelib import WareCallback
+from warelib.decorators import reset as ware_reset
+
+from eoclib.callbacks import EOCWareMain, EOCWareSetup, EOCWareThumbnail
+from eoclib.types import WareData
 
 
-def setup(
-    *args, **kwargs
-) -> Callable[[Callable[[WareConfig, WareData], None]], EOCWareSetup]:
-    def callback(func: Callable[[WareConfig, WareData], None]) -> EOCWareSetup:
-        return EOCWareSetup(func)
-
-    return callback
+def setup(func: Callable[[WareData], None]) -> EOCWareSetup:
+    return EOCWareSetup(func)
 
 
 def thumbnail(
-    *args, **kwargs
-) -> Callable[
-    [Callable[..., Generator[None, tuple[pygame.Surface, list[pygame.Event]], None]]],
-    EOCWareThumbnail,
-]:
-    def callback(
-        func: Callable[
-            ..., Generator[None, tuple[pygame.Surface, list[pygame.Event]], None]
-        ],
-    ) -> EOCWareThumbnail:
-        return EOCWareThumbnail(func)
-
-    return callback
+    func: Callable[
+        ..., Generator[None, tuple[pygame.Surface, list[pygame.Event]], None]
+    ],
+) -> EOCWareThumbnail:
+    return EOCWareThumbnail(func)
 
 
 def mainloop(
-    *args, **kwargs
-) -> Callable[
-    [Callable[..., Generator[None, tuple[pygame.Surface, list[pygame.Event]], None]]],
-    EOCWareMainloop,
-]:
-    def callback(
-        func: Callable[
-            ..., Generator[None, tuple[pygame.Surface, list[pygame.Event]], None]
-        ],
-    ) -> EOCWareMainloop:
-        return EOCWareMainloop(func)
+    func: Callable[
+        ..., Generator[None, tuple[pygame.Surface, list[pygame.Event]], None]
+    ],
+) -> EOCWareMain:
+    return EOCWareMain(func)
 
-    return callback
+
+def reset(func: Callable[[], None]) -> WareCallback[[], None]:
+    return ware_reset(func)
